Move (most) Debugger Attributes to shared partition
